Remove USB host mode support in kernel
This makes it so that the USB port *only* works in gadget mode. It fixes
an issue where the USB connection didn't work to my laptop. Presumably,
the USB stack thinks that the port should be run in host mode. However,
it was unclear why everything worked in Raspbian.

See issue #10 for more details.